refactor: 헤더 공통화와 스크롤 배치 정리
