+++ src/app/(protected)/ideas/page.tsx
@@ -19,23 +19,23 @@ export default function IdeasPage() {
     const openSidebar = useLayoutStore((state) => state.openSidebar)
 
     return (
-        <div className="flex h-full w-full flex-col bg-bg-0">
-            <Header
-                title="트렌드 · 아이디어"
-                leadingClassName="desktop:hidden"
-                leading={
-                    <button
-                        type="button"
-                        onClick={openSidebar}
-                        className="-ml-1 flex size-8 items-center justify-center text-icon-primary transition-colors hover:text-text-primary"
-                        aria-label="메뉴 열기"
-                    >
-                        <MenuIcon />
-                    </button>
-                }
-            />
-
+        <div className="flex h-full w-full flex-col bg-bg-0 desktop:pt-3">
             <Scroll as="main" className="flex-1">
+                <Header
+                    title="트렌드 · 아이디어"
+                    leadingClassName="desktop:hidden"
+                    showDivider={false}
+                    leading={
+                        <button
+                            type="button"
+                            onClick={openSidebar}
+                            className="-ml-1 flex size-8 items-center justify-center text-icon-primary transition-colors hover:text-text-primary"
+                            aria-label="메뉴 열기"
+                        >
+                            <MenuIcon />
+                        </button>
+                    }
+                />
                 <div className="flex flex-col gap-8 pb-16">
                     <PageContent className="flex flex-col items-center gap-3.5">
                         <TrendKeyword />
