+++ src/app/(protected)/settings/page.tsx
@@ -1,14 +1,16 @@
 'use client'
 
 import { useState } from 'react'
+import MenuIcon from '@/assets/icons/menu.svg'
 import Scroll from '@/components/Scroll'
+import Header from '@/components/layout/Header'
 import PageContent from '@/components/layout/PageContent'
+import { useLayoutStore } from '@/stores/layoutStore'
 import ActionRow from './_components/ActionRow'
 import EditableTextField from './_components/EditableTextField'
 import NotificationRow from './_components/NotificationRow'
 import ProfileField from './_components/ProfileField'
 import SectionDivider from './_components/SectionDivider'
-import SettingHeader from './_components/SettingHeader'
 import SettingsProfileImage from './_components/SettingsProfileImage'
 
 const channel = {
@@ -18,16 +20,30 @@ const channel = {
 }
 
 export default function SettingsPage() {
+    const openSidebar = useLayoutStore((state) => state.openSidebar)
     const [emailNotifications, setEmailNotifications] = useState({
         dailyRecommendation: false,
         marketing: true,
     })
 
     return (
-        <div className="flex h-full w-full flex-col bg-bg-0">
-            <SettingHeader />
-
+        <div className="flex h-full w-full flex-col bg-bg-0 desktop:pt-3">
             <Scroll as="main" className="flex-1">
+                <Header
+                    title="설정"
+                    leadingClassName="desktop:hidden"
+                    showDivider={false}
+                    leading={
+                        <button
+                            type="button"
+                            onClick={openSidebar}
+                            className="-ml-1 flex size-8 items-center justify-center text-icon-primary transition-colors hover:text-text-primary"
+                            aria-label="메뉴 열기"
+                        >
+                            <MenuIcon />
+                        </button>
+                    }
+                />
                 <section className="flex flex-col gap-8 pb-8">
                     <PageContent className="flex flex-col gap-[22px] pt-[17px] desktop:pt-0">
                         <SettingsProfileImage channelName={channel.name} />
